Authentication

+++ Controllers/UserController.cs
@@ -1,10 +1,13 @@
 using BikeServiceAPI.Models.DTOs;
 using BikeServiceAPI.Services;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BikeServiceAPI.Controllers;
 
 [ApiController, Route("[controller]")]
+[Authorize]
 public class UserController : ControllerBase
 {
     private readonly IUserService _userService;
@@ -15,12 +18,14 @@ public class UserController : ControllerBase
     }
 
     [HttpGet]
+    [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme, Policy = "Admin")]
     public async Task<List<UserDto>> GetAllUsers()
     {
         return await _userService.GetAllUsers();
     }
 
     [HttpGet("{id}")]
+    [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme, Policy = "Admin")]
     public async Task<IActionResult> GetUserById(long id)
     {
         try
@@ -34,6 +39,7 @@ public class UserController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme, Policy = "Admin")]
     public async Task<IActionResult> AddUser([FromBody] UserDto userDto)
     {
         await _userService.AddUser(userDto);
@@ -41,6 +47,7 @@ public class UserController : ControllerBase
     }
 
     [HttpPut]
+    [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
     public async Task<IActionResult> UpdateUser([FromBody] UserDto userDto)
     {
         try
@@ -54,9 +61,19 @@ public class UserController : ControllerBase
     }
 
     [HttpDelete("{id}")]
+    [Authorize(AuthenticationSchemes = CookieAuthenticationDefaults.AuthenticationScheme)]
     public async Task<IActionResult> DeleteUser(long id)
     {
-        try
+        var databaseIdClaim = User.Claims.FirstOrDefault(c => c.Type == "DatabaseID");
+        string databaseId = "";
+        if (databaseIdClaim != null)
+        {
+            databaseId = databaseIdClaim.Value;
+        }
+        
+        if (User.IsInRole("Admin") || databaseId == id.ToString())
+        {
+            try
         {
             return Ok(await _userService.DeleteUser(id));
         }
@@ -64,5 +81,7 @@ public class UserController : ControllerBase
         {
             return BadRequest(e.StackTrace);
         }
+        }
+        return BadRequest("You have no authority to delete this user.");
     }
 }
